% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{apply_link}
\alias{apply_link}
\title{Apply link functions to parameters}
\usage{
apply_link(par, linklist, type = c("linkfun", "linkinv", "mu.eta"))
}
\arguments{
\item{par}{vector of parameter values}

\item{linklist}{list containing \code{linkfun}, \code{linkinv}, and \code{mu.eta} for each link}

\item{type}{string specifying which function should be applied}
}
\description{
Apply link functions to parameters
}
\seealso{
\code{\link{make.link}}
}
