% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blowfly}
\alias{blowfly}
\title{Nicholson's blowfly data}
\format{
A data frame containing 361 rows comprising:

\describe{
 \item{eggs}{number of eggs}
 \item{nonemerging}{?}
 \item{emerging}{?}
 \item{deaths}{?}
 \item{total}{?}
}
}
\usage{
blowfly
}
\description{
...
}
\keyword{datasets}
