% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bombay}
\alias{bombay}
\title{Data from 1905 plague outbreak in Mumbai (formerly called Bombay)}
\format{
A data frame with 31 rows comprising:
\describe{
\item{week}{week}
\item{mort}{mortality}
}
}
\usage{
bombay
}
\description{
Plague death reports ...
}
\keyword{datasets}
