% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{check_link}
\alias{check_link}
\title{Check link functions}
\usage{
check_link(model, link)
}
\arguments{
\item{model}{odemodel object}

\item{link}{named vector specifying link functions}
}
\description{
Check link functions for model parameters -- for \code{\link{fitode}}
internal usage
}
\keyword{internal}
