% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{coef,fitode-method}
\alias{coef,fitode-method}
\title{Extract parameter of a fit}
\usage{
\S4method{coef}{fitode}(object, type = c("original", "fitted"))
}
\arguments{
\item{object}{fitode object}

\item{type}{type of parameter to be returned}
}
\description{
Extract parameter of a fit
}

