% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{coef,fitodeMCMC-method}
\alias{coef,fitodeMCMC-method}
\title{Extract model coefficients from fitodeMCMC objects}
\usage{
\S4method{coef}{fitodeMCMC}(object)
}
\arguments{
\item{object}{fitodeMCMC object}
}
\value{
The estimated median coefficients of the fitodeMCMC object
}
\description{
Extracts estimated parameters (median of the marginal posterior distributions)
}
