% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-models.R
\docType{class}
\name{de-model-class}
\alias{de-model-class}
\title{Class representing ode models}
\description{
Class representing ode models
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{model}}{list of formulas specifying ode model}

\item{\code{grad}}{ode gradients}

\item{\code{state}}{state variables}

\item{\code{par}}{parameters}

\item{\code{js}}{Jacobian with respect to its states}

\item{\code{jp}}{Jacobian with repsect to its parameters}
}}

