% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{dfun}
\alias{dfun}
\title{Taylor expansion of digamma(a+b) for a>>b}
\usage{
dfun(x, y, mag = 1e+08)
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}

\item{mag}{cutoff magnitude for switching approximations}
}
\value{
numeric
}
\description{
Taylor expansion of digamma(a+b) for a>>b
}
\keyword{internal}
