% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{dfun2}
\alias{dfun2}
\title{Taylor expansion of trigamma(a+b) (?) for a>>b}
\usage{
dfun2(x, y, mag = 1e+08, focal = "x")
}
\arguments{
\item{x}{first argument}

\item{y}{second argument}

\item{mag}{cutoff magnitude for switching approximations}
}
\value{
numeric
}
\description{
Taylor expansion of trigamma(a+b) (?) for a>>b
}
\keyword{internal}
