% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\docType{methods}
\name{Eval,loglik.ode-method}
\alias{Eval,loglik.ode-method}
\title{Evaluate the log-likelihood model}
\usage{
\S4method{Eval}{loglik.ode}(object, observation, mean, par = NULL, ...)
}
\arguments{
\item{object}{loglik.ode object}

\item{observation}{observations}

\item{mean}{mean values}

\item{par}{additional parameters}

\item{...}{other values if required}
}
\value{
numeric
}
\description{
Evaluate the log-likelihood model
}
\keyword{internal}
