% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Eval}
\alias{Eval}
\title{S4 generic for evaluate an object}
\usage{
Eval(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\description{
S4 generic for evaluate an object
}
\keyword{internal}
