% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R
\docType{class}
\name{fitode-class}
\alias{fitode}
\alias{fitode-class}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation
}
\seealso{
\code{\link{mle2-class}}
}

