% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\docType{methods}
\name{fitode}
\alias{fitode}
\title{fit ode}
\usage{
\S4method{initialize}{fitode}(.Object, formula, start, model,
  loglik = select_model("gaussian"), data, tcol = "times", links,
  control = list(maxit = 1e+05), ode.opts = list(method = "rk4", hini =
  0.1), debug = FALSE)
}
\arguments{
\item{formula}{formula specifing observation variable and mean}

\item{start}{starting values for optimization}

\item{model}{ode model}

\item{loglik}{log liklihood model}

\item{data}{data frame with time column and observation column}

\item{tcol}{time column}

\item{links}{named vector or list of link functions for ode parameters}

\item{control}{see optim}

\item{ode.opts}{options for ode integration. See ode}

\item{debug}{print debugging output?}
}
\description{
fit ode
}

