% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC.R
\name{fitodeMCMC}
\alias{fitodeMCMC}
\title{Fit ordinary differential equations model using MCMC}
\usage{
fitodeMCMC(model, data, start, tcol = "times", vcov, prior = list(),
  chains = 1, mcmc = 2000, burnin = 1000, thin = 1,
  prior.only = FALSE, link, fixed = list(), control = list(maxit =
  1e+05), solver.opts = list(method = "rk4"), solver = ode,
  skip.hessian = FALSE, force.hessian = FALSE, use.ginv = TRUE, ...)
}
\arguments{
\item{model}{ode model}

\item{data}{data frame with time column and observation column}

\item{start}{named vector of starting parameter values}

\item{tcol}{time column}

\item{prior.only}{(logical) sample from prior distribution only?}

\item{link}{named vector or list of link functions for model parameters}

\item{fixed}{named vector or list of model parameters to fix}

\item{solver.opts}{options for ode integration. See \code{\link{ode}}}

\item{solver}{ode solver}

\item{debug}{print debugging output?}
}
\description{
Fit ordinary differential equations model using MCMC
}
