% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{fixpar}
\alias{fixpar}
\title{Fix parameters of an ODE model}
\usage{
fixpar(model, fixed)
}
\arguments{
\item{model}{odemodel object}

\item{fixed}{named vector or list of model parameters to fix}
}
\description{
Fix parameters of an ODE model to a constant value by transforming
the model (using Transform() function)
}
\keyword{internal}
