% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-models.R
\docType{methods}
\name{grad,de-model-method}
\alias{grad,de-model-method}
\title{Evaluate the gradients of a model}
\usage{
\S4method{grad}{`de-model`}(object, state, par)
}
\arguments{
\item{object}{object to be evaluated}

\item{state}{state}

\item{par}{additional parameters}
}
\description{
Evaluate the gradients of a model
}

