% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\docType{methods}
\name{grad,loglik.ode-method}
\alias{grad,loglik.ode-method}
\title{Evaluate the gradient of a likelihood model}
\usage{
\S4method{grad}{loglik.ode}(object, observation, mean, par, ...)
}
\arguments{
\item{object}{loglik.ode object}

\item{observation}{observations}

\item{mean}{mean values}

\item{par}{additional parameters}

\item{...}{other values if required}
}
\value{
a list with each element as a partial derivative values
}
\description{
Evaluate the gradient of a likelihood model
}
\keyword{internal}
