% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{grad,odemodel-method}
\alias{grad,odemodel-method}
\title{Evaluate the gradients of a model}
\usage{
\S4method{grad}{odemodel}(object, state, par)
}
\arguments{
\item{object}{odemodel object}

\item{state}{state}

\item{par}{parameter values}
}
\description{
Evaluate the gradients of a model
}
\keyword{internal}
