% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{grad}
\alias{grad}
\title{S4 generic for computing a gradient}
\usage{
grad(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\description{
S4 generic for computing a gradient
}
\keyword{internal}
