% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{hessian}
\alias{hessian}
\title{S4 generic for computing a hessian}
\usage{
hessian(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\value{
The hessian of the object.
}
\description{
S4 generic for computing a hessian
}
\keyword{internal}
