% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-models.R
\docType{methods}
\name{initialize,de-model-method}
\alias{initialize,de-model-method}
\title{the initializer for de-model}
\usage{
\S4method{initialize}{`de-model`}(.Object, name, model, state, par,
  keep_jacobian = TRUE)
}
\arguments{
\item{.Object}{object}
}
\description{
the initializer for de-model
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{model}}{ode model}

\item{\code{state}}{state variables}

\item{\code{par}}{parameters}

\item{\code{keep_jacobian}}{(logical) maintain the Jacobian as part of the model}
}}
\examples{
SI_model <- new("de-model",
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    state = c("S", "I"),
    par= c("beta", "gamma", "N")
)
}

