% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\docType{methods}
\name{initialize,loglik.ode-method}
\alias{initialize,loglik.ode-method}
\title{The initializer for loglik.ode}
\usage{
\S4method{initialize}{loglik.ode}(
  .Object,
  name,
  model,
  observation = "X",
  mean,
  par = NULL,
  keep_grad = TRUE
)
}
\arguments{
\item{.Object}{object}

\item{name}{name of the distribution}

\item{model}{formula specifying the disstribution}

\item{observation}{observation variable name}

\item{mean}{mean variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}
}
\value{
An object of class ``loglik.ode'' as described in \code{\link{loglik.ode-class}}.
}
\description{
The initializer for loglik.ode
}
\keyword{internal}
