% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{initialize,model.ode-method}
\alias{initialize,model.ode-method}
\title{the initializer for model.ode}
\usage{
\S4method{initialize}{model.ode}(.Object, name, model, initial, par,
  keep_sensitivity = TRUE)
}
\arguments{
\item{.Object}{object}
}
\description{
the initializer for model.ode
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{model}}{ode model}

\item{\code{initial}}{initial values}

\item{\code{par}}{parameters}

\item{\code{keep_sensitivity}}{(logical) maintain the Jacobian as part of the model}
}}
\examples{
SI_model <- new("model.ode",
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    initial = list(
        S ~ N * (1 - i0),
        I ~ N * i0
    ),
    par= c("beta", "gamma", "N", "i0")
)
}

