% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{methods}
\name{initialize,prior.ode-method}
\alias{initialize,prior.ode-method}
\title{the initializer for prior.ode}
\usage{
\S4method{initialize}{prior.ode}(
  .Object,
  name,
  model,
  observation = "X",
  par = NULL,
  keep_grad = TRUE
)
}
\arguments{
\item{.Object}{object}

\item{name}{name of the distribution}

\item{model}{the formula specifying the model}

\item{observation}{observation variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}
}
\value{
An object of class ``prior.ode'' as described in \code{\link{prior.ode-class}}.
}
\description{
the initializer for prior.ode
}
\keyword{internal}
