% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{methods}
\name{initialize,solution.ode-method}
\alias{initialize,solution.ode-method}
\title{the initializer for model.ode}
\usage{
\S4method{initialize}{solution.ode}(.Object, model, result,
  keep_sensitivity = TRUE)
}
\arguments{
\item{.Object}{object}
}
\description{
the initializer for model.ode
}
\section{Slots}{

\describe{
\item{\code{model}}{ode model}

\item{\code{result}}{deSolve result}
}}

