% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-models.R
\docType{methods}
\name{jacobian,de-model-method}
\alias{jacobian,de-model-method}
\title{Evaluate the jacobian of the gradients}
\usage{
\S4method{jacobian}{`de-model`}(object, state, par, type = c("state", "par"))
}
\arguments{
\item{object}{object to be evaluated}

\item{state}{state}

\item{par}{additional parameters}

\item{type}{state of par?}
}
\description{
Evaluate the jacobian of the gradients
}

