% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{jacobian,model.ode-method}
\alias{jacobian,model.ode-method}
\title{Evaluate the jacobian of the gradients}
\usage{
\S4method{jacobian}{model.ode}(object, state, par, type = c("initial",
  "state", "par"))
}
\arguments{
\item{object}{object to be evaluated}

\item{state}{state}

\item{par}{additional parameters}

\item{type}{state of par?}
}
\description{
Evaluate the jacobian of the gradients
}
