% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{jacobian,odemodel-method}
\alias{jacobian,odemodel-method}
\title{Evaluate the jacobian of the gradients}
\usage{
\S4method{jacobian}{odemodel}(object, state, par, type = c("initial", "state", "par"))
}
\arguments{
\item{object}{odemodel object}

\item{state}{state}

\item{par}{parameter values}

\item{type}{state of par?}
}
\description{
Evaluate the jacobian of the gradients
}
\keyword{internal}
