% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{jacobian}
\alias{jacobian}
\title{S4 generic for computing a jacobian}
\usage{
jacobian(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\description{
S4 generic for computing a jacobian
}
