% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{logLik,fitode-method}
\alias{logLik,fitode-method}
\title{Extract log-likelihood of a fit}
\usage{
\S4method{logLik}{fitode}(object)
}
\arguments{
\item{object}{fitode object}
}
\description{
Extract log-likelihood of a fit
}
