% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{make_prior}
\alias{make_prior}
\title{Make a list containing log prior density and its gradient}
\usage{
make_prior(model, link, prior, prior.density = TRUE, keep_grad = TRUE)
}
\arguments{
\item{model}{model.ode object}

\item{link}{link}

\item{prior}{list of formulas}

\item{prior.density}{(logical) does this represent a probability density?}

\item{keep_grad}{(logical) keep gradient?}
}
\description{
Make a list containing log prior density and its gradient
}
\keyword{internal}
