% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R
\docType{class}
\name{model.ode-class}
\alias{model.ode-class}
\title{Class representing ode models}
\description{
Class representing ode models
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{gfun}}{gradient function}

\item{\code{initial}}{list of expressions representing the initial values}

\item{\code{state}}{state variables}

\item{\code{par}}{parameters}

\item{\code{jacobian.initial}}{Jacobian of initial values with respect to its parameters}

\item{\code{jacobian.state}}{Jacobian with respect to its states}

\item{\code{jacobian.par}}{Jacobian with repsect to its parameters}

\item{\code{keep_sensitivity}}{(logical) keep sensitivity equations?}
}}

