% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{ode.sensitivity}
\alias{ode.sensitivity}
\title{Calculate sensitivity of the likelihood function with respect to the parameters}
\usage{
ode.sensitivity(parms, formula, model, loglik, observation, times = NULL,
  ode.opts = list(method = "rk4", hini = 0.1))
}
\arguments{
\item{parms}{ode/likelihood function parameters}

\item{formula}{formula specifing observation variable and mean}
}
\description{
Calculate sensitivity of the likelihood function with respect to the parameters
}

