% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{ode.solve}
\alias{ode.solve}
\title{solve ode models}
\usage{
ode.solve(model, times, parms, y, ode.opts = list(method = "lsoda"),
  keep_sensitivity = TRUE)
}
\arguments{
\item{model}{model.ode object}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{ode.opts}{options for ode integration}

\item{keep_sensitivity}{keep sensitivity equations}
}
\description{
solve ode models
}

