% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/model.R
\docType{class}
\name{odemodel-class}
\alias{odemodel-class}
\alias{odemodel}
\title{Class "odemodel" representing ode models}
\usage{
\S4method{initialize}{odemodel}(.Object, name, model, observation, initial,
  par, link, diffnames, keep_sensitivity = TRUE)

odemodel(...)
}
\arguments{
\item{.Object}{object}

\item{name}{name of the model}

\item{model}{ode model}

\item{observation}{observation model}

\item{initial}{initial values}

\item{par}{model parameters}

\item{link}{link functions for parameters (log links are used as default)}

\item{diffnames}{optional character vector specifying the names of a variable for which the consecutive difference needs to be calculated}

\item{keep_sensitivity}{(logical) maintain the Jacobian as a part of the model object?}
}
\description{
Class "odemodel" representing ode models

Constructor method of "odemodel" class

Wrapper function odemodel
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{gfun}}{gradient function}

\item{\code{grad}}{list of gradients}

\item{\code{observation}}{list of observation models}

\item{\code{initial}}{list of expressions representing the initial values}

\item{\code{state}}{state variables}

\item{\code{par}}{parameters}

\item{\code{link}}{link functions for parameters (log links are used as default)}

\item{\code{diffnames}}{character vector specifying the names of a variable for which the consecutive difference needs to be calculated}

\item{\code{jacobian.initial}}{Jacobian of initial values with respect to its parameters}

\item{\code{jacobian.state}}{Jacobian with respect to its states}

\item{\code{jacobian.par}}{Jacobian with repsect to its parameters}

\item{\code{loglik}}{list of log-likelihood functions}

\item{\code{expr}}{expressions for true trajectories}

\item{\code{expr.sensitivity}}{sensitivity of the expressions with respect to state variables and parameters}

\item{\code{keep_sensitivity}}{(logical) keep sensitivity equations?}
}}

\examples{
SI_model <- odemodel(
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    observation = list(
        susceptible ~ dnorm(mean=S, sd=sigma1),
        infected ~ dnorm(mean=I, sd=sigma2)
    ),
    initial = list(
        S ~ N * (1 - i0),
        I ~ N * i0
    ),
    par = c("beta", "gamma", "N", "i0", "sigma1", "sigma2"),
    link = c(i0="logit")
)
}
\keyword{internal}
