% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phila1918}
\alias{phila1918}
\title{1918 Philadelphia flu data set}
\format{
A data frame with 122 rows comprising:
\describe{
\item{date}{date}
\item{mort}{mortality}
}
}
\usage{
phila1918
}
\description{
1918 Philadelphia flu data set
}
\keyword{datasets}
