% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{plot,fitode,missing-method}
\alias{plot,fitode,missing-method}
\alias{plot,fitode-method}
\title{Plot a fitode object}
\usage{
\S4method{plot}{fitode,missing}(x, level, method = c("delta", "mvrnorm",
  "wmvrnorm"), main, xlim, ylim, xlab, ylab, add = FALSE,
  col.traj = "black", lty.traj = 1, col.conf = "red", lty.conf = 4,
  nsim = 1000, ...)
}
\arguments{
\item{x}{fitode object}

\item{level}{the confidence level required}

\item{method}{confidence interval method}

\item{main}{an overall title for the plot}

\item{xlim}{the x limit of the plot}

\item{ylim}{the y limit of the plot}

\item{xlab}{a label for the x axis}

\item{ylab}{a label for the y axis}

\item{add}{(logical) add to an existing plot?}

\item{col.traj}{colour of the estimated trajectory}

\item{lty.traj}{line type of the estimated trajectory}

\item{col.conf}{colour of the confidence intervals}

\item{lty.conf}{line type of the confidence intervals}

\item{nsim}{number of simulations for mvrnorm, wmvrnorm methods}

\item{...}{additional arguments to be passed on to the plot function}
}
\description{
Plot a fitode object
}

