% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_internal}
\alias{plot_internal}
\title{Internal function for plotting methods}
\usage{
plot_internal(
  pred,
  data,
  onepage = TRUE,
  xlim,
  ylim,
  xlabs,
  ylabs,
  col.traj = "black",
  lty.traj = 1,
  col.conf = "black",
  lty.conf = 4,
  add = FALSE,
  ...
)
}
\arguments{
\item{pred}{prediction objects}

\item{data}{observed data}

\item{onepage}{(logical) print all figures on one page?}

\item{xlim}{x coordinates range}

\item{ylim}{y coordinates range}

\item{xlabs}{a label for the x axis}

\item{ylabs}{a label for the y axis}

\item{col.traj}{colour of the estimated trajectory}

\item{lty.traj}{line type of the estimated trajectory}

\item{col.conf}{colour of the confidence intervals}

\item{lty.conf}{line type of the confidence intervals}

\item{add}{add to another plot?}

\item{...}{additional arguments to be passed on to the plot function}
}
\description{
Internal function for plotting methods
}
