% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{predict,fitodeMCMC-method}
\alias{predict,fitodeMCMC-method}
\title{Prediction function for fitodeMCMC objects}
\usage{
\S4method{predict}{fitodeMCMC}(object, level, times, simplify = TRUE)
}
\arguments{
\item{object}{fitodeMCMC object}

\item{level}{the credible level required}

\item{times}{time vector to predict over. Default is set to the time frame of the data.}

\item{simplify}{(logical) simplify output to return estimated trajectories and their
credible intervals? If \code{simplify=FALSE}, all posterior trajectories will be returned}
}
\value{
Estimated trajectories and their credible intervals of the fitodeMCMC object
}
\description{
Computes estimated trajectories and their credible intervals.
The estimated trajectories are obtained by taking the median trajectories
from the posterior samples.
}
