% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R
\docType{class}
\name{prior.ode-class}
\alias{prior.ode-class}
\title{Class representing prior models used to fit ode models}
\description{
Class representing prior models used to fit ode models
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the distribution}

\item{\code{expr}}{an expression specifying the model}

\item{\code{observation}}{observation variable name}

\item{\code{par}}{additional parameter names}

\item{\code{keep_grad}}{keep gradient?}

\item{\code{grad}}{the gradient with respect to the parameters}
}}

