% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{profile,fitode-method}
\alias{profile,fitode-method}
\title{Profile object}
\usage{
\S4method{profile}{fitode}(fitted, trace = FALSE)
}
\arguments{
\item{fitted}{fitted model object}
}
\description{
Profile object
}

