% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{select_model}
\alias{select_model}
\title{Select likelihood model}
\usage{
select_model(dist = c("dnorm", "dnorm2", "dpois", "dnbinom", "dnbinom1"))
}
\arguments{
\item{dist}{conditional distribution of reported data}
}
\description{
Select likelihood model
}
