% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{select_prior}
\alias{select_prior}
\title{Select a prior model}
\usage{
select_prior(
  family = c("dnorm", "dgamma", "dbeta", "dlnorm"),
  link = c("identity", "log", "logit"),
  prior.density = TRUE,
  keep_grad = TRUE
)
}
\arguments{
\item{family}{prior distribution type}

\item{link}{link}

\item{prior.density}{(logical) keep the Jacobian of transformations?}

\item{keep_grad}{(logical) keep gradients?}
}
\description{
Select a prior model
}
\keyword{internal}
