% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{set_link}
\alias{set_link}
\title{Set up link functions for model parameters}
\usage{
set_link(link, modelpar)
}
\arguments{
\item{link}{named list or vector of strings specifying link functions}

\item{model}{model.ode object}

\item{loglik}{loglik.ode object}
}
\value{
list of strings specifying link functions
}
\description{
Set up link functions for model parameters
}
\seealso{
\code{\link{make.link}}
}
