% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{set_links}
\alias{set_links}
\title{Set up link functions for ode/loglik parameters}
\usage{
set_links(links, model, loglik)
}
\arguments{
\item{model}{model.ode object}

\item{loglik}{loglik.ode object}

\item{link}{named list or vector of strings specifying link functions}
}
\value{
list of strings specifying link functions
}
\description{
Set up link functions for ode/loglik parameters
}
\seealso{
\code{\link{make.link}}
}

