% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{show,model.ode-method}
\alias{show,model.ode-method}
\title{Show the model}
\usage{
\S4method{show}{model.ode}(object)
}
\arguments{
\item{object}{model.ode object}
}
\description{
Show the model
}
