% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SierraLeone2014}
\alias{SierraLeone2014}
\title{Data from 2014 Sierra Leone Ebola epidemic}
\format{
A data frame with 67 rows comprising:
\describe{
\item{times}{decimal dates (year + fraction of year)}
\item{confirmed}{confirmed cases}
}
}
\usage{
SierraLeone2014
}
\description{
Ebola case reports ...
}
\keyword{datasets}
