% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\docType{methods}
\name{simulate,fitode-method}
\alias{simulate,fitode-method}
\title{simulate fitode objects}
\usage{
\S4method{simulate}{fitode}(object, nsim = 1, seed = NULL, times, parms, y, observation = TRUE)
}
\arguments{
\item{object}{fitode object}

\item{nsim}{number of simulations}

\item{seed}{random-number seed}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{observation}{(logical) propagate observation error?}
}
\value{
The numerical simulation of the object
}
\description{
simulate fitode objects
}
