% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\docType{methods}
\name{simulate,model.ode-method}
\alias{simulate,model.ode-method}
\title{simulate model objects}
\usage{
\S4method{simulate}{model.ode}(object, nsim = 1, seed = NULL, times,
  parms, y, solver.opts = list(method = "rk4"), solver = ode,
  observation = TRUE)
}
\arguments{
\item{object}{model.ode object}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{solver.opts}{options for ode solver}

\item{solver}{ode solver (must take y, times, func, and parms arguments)}

\item{observation}{(logical) propagate observation error?}
}
\description{
simulate model objects
}
