% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\docType{methods}
\name{simulate,odemodel-method}
\alias{simulate,odemodel-method}
\title{simulate model objects}
\usage{
\S4method{simulate}{odemodel}(
  object,
  nsim = 1,
  seed = NULL,
  times,
  parms,
  y,
  solver.opts = list(method = "rk4"),
  solver = ode,
  observation = TRUE
)
}
\arguments{
\item{object}{odemodel object}

\item{nsim}{number of simulations}

\item{seed}{random-number seed}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{solver.opts}{options for ode solver}

\item{solver}{ode solver (must take y, times, func, and parms arguments)}

\item{observation}{(logical) propagate observation error?}
}
\value{
The numerical simulation of the object
}
\description{
simulate model objects
}
