% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_internal}
\alias{simulate_internal}
\title{Internal function for simulation models}
\usage{
simulate_internal(
  model,
  times,
  parms,
  y,
  solver.opts = list(method = "rk4"),
  solver = ode,
  observation = TRUE,
  nsim = 1,
  seed = NULL
)
}
\arguments{
\item{model}{odemodel object}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{solver.opts}{options for ode solver}

\item{solver}{ode solver (must take y, times, func, and parms arguments)}

\item{observation}{(logical) propagate observation error?}

\item{nsim}{number of simulations}

\item{seed}{seed}
}
\description{
Simulates deterministic trajectories and propagates observation error
}
