% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\docType{class}
\name{solution.ode-class}
\alias{solution.ode-class}
\title{Class "solution.ode".
Result of solving ode modeld with/without sensitivity equations}
\description{
Class "solution.ode".
Result of solving ode modeld with/without sensitivity equations
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{y}}{initial values}

\item{\code{times}}{time vector}

\item{\code{model}}{ode model}

\item{\code{parms}}{parameters of the solution}

\item{\code{solution}}{solution of the model}

\item{\code{sensitivity}}{partial derivative of each state variable with respect to the parameters}
}}

\keyword{internal}
