% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methos}
\name{stdEr,fitode-method}
\alias{stdEr,fitode-method}
\title{Calculate standard error}
\usage{
\S4method{stdEr}{fitode}(x, scale = c("original", "fitted"))
}
\description{
Calculate standard error
}

