% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{stdEr,fitodeMCMC-method}
\alias{stdEr,fitodeMCMC-method}
\title{Extract standard error}
\usage{
\S4method{stdEr}{fitodeMCMC}(x)
}
\arguments{
\item{x}{fitodeMCMC object}
}
\description{
Calculates standard error by taking the square root of the diagonal matrix
}
