% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{summary,fitode-method}
\alias{summary,fitode-method}
\title{Summarize \code{fitode} object}
\usage{
\S4method{summary}{fitode}(object)
}
\arguments{
\item{object}{fitode object}
}
\description{
Summarize \code{fitode} object
}
