% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{summary,fitodeMCMC-method}
\alias{summary,fitodeMCMC-method}
\title{Summarize \code{fitodeMCMC} object}
\usage{
\S4method{summary}{fitodeMCMC}(object)
}
\arguments{
\item{object}{fitodeMCMC object}
}
\description{
Summarize \code{fitodeMCMC} object;
returns estimate, standard error, confidence intervals, effective sample sizes, and gelman-rubin diagnostic
}
\seealso{
\code{\link{effectiveSize}} \code{\link{gelman.diag}}
}
