% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{to.formula}
\alias{to.formula}
\title{create formula from a character and a call}
\usage{
to.formula(left, right)
}
\arguments{
\item{left}{(character) lhs of the formula}

\item{right}{(call) rhs of the formula}
}
\description{
create formula from a character and a call
}
\examples{
to.formula("beta", substitute(exp(log.beta)))
}

