% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\docType{methods}
\name{Transform,loglik.ode-method}
\alias{Transform,loglik.ode-method}
\title{Transform the model}
\usage{
\S4method{Transform}{loglik.ode}(
  object,
  transforms = NULL,
  name,
  observation = "X",
  mean,
  par,
  keep_grad = TRUE
)
}
\arguments{
\item{object}{loglik.ode object}

\item{transforms}{list of formulas specifying transformations}

\item{name}{name of the log-likelihood model}

\item{observation}{observation variable name}

\item{mean}{mean variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}
}
\value{
loglik.ode object
}
\description{
Transform the model
}
\keyword{internal}
