% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{Transform,odemodel-method}
\alias{Transform,odemodel-method}
\title{Transform the model}
\usage{
\S4method{Transform}{odemodel}(object, transforms = NULL, par,
  keep_sensitivity)
}
\arguments{
\item{object}{odemodel object}

\item{transforms}{list of formulas specifying transformations}

\item{par}{model parameters}

\item{keep_sensitivity}{(logical) maintain the Jacobian as part of the model}
}
\description{
Transform the model
}
\keyword{internal}
