% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{methods}
\name{Transform,prior.ode-method}
\alias{Transform,prior.ode-method}
\title{Transform the prior model}
\usage{
\S4method{Transform}{prior.ode}(
  object,
  transforms = NULL,
  name,
  observation = "X",
  par,
  keep_grad = TRUE
)
}
\arguments{
\item{object}{object}

\item{transforms}{list of formulas specifying transformations}

\item{name}{name of the log-likelihood model}

\item{observation}{observation variable name}

\item{par}{additional parameter names}

\item{keep_grad}{maintain the gradient as part of the model}
}
\value{
An object of class ``prior.ode'' as described in \code{\link{prior.ode-class}}.
}
\description{
Transform the prior model
}
\keyword{internal}
