% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Transform}
\alias{Transform}
\title{S4 generic for transforming an object}
\usage{
Transform(object, ...)
}
\arguments{
\item{object}{an \code{R} object}

\item{...}{further arguments passed to methods}
}
\value{
The result of transforming the object.
}
\description{
S4 generic for transforming an object
}
\keyword{internal}
