% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transpar}
\alias{transpar}
\title{transform parameters}
\usage{
transpar(parms, transform, inverse)
}
\arguments{
\item{parms}{numeric vector containing parameters}

\item{transform}{list of transformations specified as a formula}

\item{inverse}{list of inverse transformations specified as formula}
}
\description{
transform parameters
}
\examples{
par <- c(a=2, b=1)
transform <- list(a~exp(log.a)+1)
inverse <- list(log.a~log(a-1))

print(newpar <- transpar(par, transform, inverse))

## direction of transformation is automatically determined
print(oldpar <- transpar(newpar, transform, inverse))
all.equal(par, oldpar)

print(oldpar2 <- transpar(newpar, inverse, transform))
all.equal(par, oldpar2)
}

