% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumorgrowth}
\alias{tumorgrowth}
\title{Tumor growth data}
\format{
A data frame containing 14 rows comprising:

\describe{
 \item{day}{}
 \item{volume}{}
}
}
\usage{
tumorgrowth
}
\description{
...
}
\keyword{datasets}
