% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update,fitode-method}
\alias{update,fitode-method}
\title{Update fitode fits}
\usage{
\S4method{update}{fitode}(object, observation, initial, par, link, ...)
}
\arguments{
\item{object}{fitode objects}

\item{observation}{observation model}

\item{initial}{initial values}

\item{par}{model parameters}

\item{link}{link functions for parameters (log links are used as default)}

\item{...}{additional arguments to be passed to fitode}
}
\value{
An object of class ``fitode'' as described in \code{\link{fitode-class}}.
}
\description{
Update fitode fits
}
