% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{vcov,fitodeMCMC-method}
\alias{vcov,fitodeMCMC-method}
\title{Extract variance-covariance matrix from fitodeMCMC objects}
\usage{
\S4method{vcov}{fitodeMCMC}(object)
}
\arguments{
\item{object}{fitodeMCMC object}
}
\description{
Calculates variance-covariance matrix from posterior samples
}
