% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{coef,fitode-method}
\alias{coef,fitode-method}
\title{Extract model coefficients from fitode objects}
\usage{
\S4method{coef}{fitode}(object, type = c("response", "links"))
}
\arguments{
\item{object}{fitode object}

\item{type}{type of coefficients. The default (\code{type=response}) is on the
response scale; this is the scale on which the model parameters are defined.
Alternatively, \code{type=link} can be used to obtain parameters on the estimated scale.}
}
\value{
The estimated coefficients of the fitode object
}
\description{
Extracts estimated parameters (either on response scales or link scales)
}
