% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{confint,fitodeMCMC-method}
\alias{confint,fitodeMCMC-method}
\title{Calculate credible intervals from fitodeMCMC objects for model parameters and their transformations}
\usage{
\S4method{confint}{fitodeMCMC}(object, parm, level = 0.95)
}
\arguments{
\item{object}{fitodeMCMC object}

\item{parm}{character vector specifying model parameters or list of formuals specifying transformations}

\item{level}{the credible level required}
}
\value{
The credible intervals of the fitodeMCMC object
}
\description{
Calculate credible intervals for model parameters and their transformations
from posterior samples.
}
