% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R
\docType{class}
\name{fitode-class}
\alias{fitode-class}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation
}
\section{Slots}{

\describe{
\item{\code{call}}{(languge) The call to \code{\link{fitode}}}

\item{\code{model}}{odemodel object}

\item{\code{data}}{the time series data}

\item{\code{coef}}{estimated parameters}

\item{\code{vcov}}{estimated variance-covariance matrix}

\item{\code{min}}{minimum negative log-likelihood}

\item{\code{mle2}}{mle2 object}

\item{\code{link}}{list of link functions for model parameters}

\item{\code{fixed}}{list of fixed parameters}

\item{\code{prior}}{list of priors}
}}

\seealso{
\code{\link{mle2-class}}
}
