% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/fitode.R
\name{fitode}
\alias{fitode}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\usage{
fitode(formula, start, model, loglik = select_model("gaussian"), data,
  method = "BFGS", optimizer = "optim", link, control = list(maxit =
  1e+05), ode.opts = list(method = "rk4"), skip.hessian = FALSE,
  use.ginv = TRUE, debug = FALSE, ...)

fitode(formula, start, model, loglik = select_model("gaussian"), data,
  method = "BFGS", optimizer = "optim", link, control = list(maxit =
  1e+05), ode.opts = list(method = "rk4"), skip.hessian = FALSE,
  use.ginv = TRUE, debug = FALSE, ...)
}
\arguments{
\item{formula}{formula specifing observation variable and the mean}

\item{start}{named vector of starting parameter values}

\item{model}{ode model}

\item{loglik}{log liklihood model}

\item{data}{data frame with time column and observation column}

\item{link}{named vector or list of link functions for ode/log-likelihood parameters}

\item{control}{see optim}

\item{ode.opts}{options for ode integration. See ode}

\item{skip.hessian}{skip hessian calculation}

\item{use.ginv}{use generalized inverse (\code{\link{ginv}}) to compute approximate vcov}

\item{debug}{print debugging output?}

\item{...}{mle2 arguments}
}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation

fit ode
}
\seealso{
\code{\link{mle2-class}}

\code{\link{mle2}}
}

