% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/fitode.R
\docType{class}
\name{fitode}
\alias{fitode}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\usage{
\S4method{initialize}{fitode}(.Object, formula, start, model,
  loglik = select_model("gaussian"), data, method = "BFGS",
  optimizer = "optim", tcol = "times", links, control = list(maxit =
  1e+05), ode.opts = list(method = "rk4", hini = 0.1), debug = FALSE)
}
\arguments{
\item{formula}{formula specifing observation variable and the mean}

\item{start}{named vector of starting parameter values}

\item{model}{ode model}

\item{loglik}{log liklihood model}

\item{data}{data frame with time column and observation column}

\item{tcol}{time column}

\item{links}{named vector or list of links for ode/log-likelihood parameters}

\item{control}{see optim}

\item{ode.opts}{options for ode integration. See ode}

\item{debug}{print debugging output?}
}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation

fit ode
}
\seealso{
\code{\link{mle2-class}}
}

