% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{fitode}
\alias{fitode}
\title{Fit ordinary differential equations model}
\usage{
fitode(model, data, start, tcol = "times", method = "BFGS",
  optimizer = "optim", link, fixed = list(), prior = list(),
  prior.density = TRUE, control = list(maxit = 1e+05),
  solver.opts = list(method = "rk4"), solver = ode,
  skip.hessian = FALSE, force.hessian = FALSE, use.ginv = TRUE, ...)
}
\arguments{
\item{model}{model.ode object}

\item{data}{data frame with time column and observation column}

\item{start}{named vector of starting parameter values}

\item{tcol}{time column}

\item{method}{optimization method}

\item{optimizer}{optimizer}

\item{link}{named vector or list of link functions for model parameters}

\item{fixed}{named vector or list of model parameters to fix}

\item{prior}{list of formulas specifying prior distributions}

\item{control}{see \code{\link{optim}}}

\item{solver.opts}{options for ode integration. See \code{\link{ode}}}

\item{solver}{ode solver}

\item{skip.hessian}{skip hessian calculation}

\item{force.hessian}{(FALSE) calculate the hessian numerically instead of taking the jacobian of the analytical gradients}

\item{use.ginv}{use generalized inverse (\code{\link{ginv}}) to compute approximate vcov}

\item{...}{mle2 arguments}

\item{prior.distribution}{(logical) should priors represent a probability distribution?}
}
\description{
Fit ordinary differential equations model
}
\seealso{
\code{\link{mle2}}
}
