% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{fitode}
\alias{fitode}
\title{fit ode}
\usage{
fitode(formula, start, model, loglik, data, tcol = "times",
  control = list(maxit = 1e+05), debug = FALSE)
}
\arguments{
\item{formula}{formula specifing observation variable and mean}

\item{data}{data frame}
}
\description{
fit ode
}

