% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/fitode.R
\name{fitode}
\alias{fitode}
\title{Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation}
\usage{
fitode(model, data, start, tcol = "times", method = "BFGS",
  optimizer = "optim", link, fixed = list(), control = list(maxit =
  1e+05), solver.opts = list(method = "rk4"), solver = ode,
  skip.hessian = FALSE, use.ginv = TRUE, debug = FALSE, ...)

fitode(model, data, start, tcol = "times", method = "BFGS",
  optimizer = "optim", link, fixed = list(), control = list(maxit =
  1e+05), solver.opts = list(method = "rk4"), solver = ode,
  skip.hessian = FALSE, use.ginv = TRUE, debug = FALSE, ...)
}
\arguments{
\item{model}{ode model}

\item{data}{data frame with time column and observation column}

\item{start}{named vector of starting parameter values}

\item{tcol}{time column}

\item{method}{optimization method}

\item{optimizer}{optimizer}

\item{link}{named vector or list of link functions for ode/log-likelihood parameters}

\item{control}{see optim}

\item{solver.opts}{options for ode integration. See ode}

\item{skip.hessian}{skip hessian calculation}

\item{use.ginv}{use generalized inverse (\code{\link{ginv}}) to compute approximate vcov}

\item{debug}{print debugging output?}

\item{...}{mle2 arguments}
}
\description{
Class "fitode".
Result of ode fitting based on Maximum Likelihood Estimation

fit ode
}
\seealso{
\code{\link{mle2-class}}

\code{\link{mle2}}
}
