% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R
\docType{class}
\name{fitodeMCMC-class}
\alias{fitodeMCMC-class}
\title{Class "fitodeMCMC".
Result of ode fitting based on Markov Chain Monte Carlo (MCMC)}
\description{
Class "fitodeMCMC".
Result of ode fitting based on Markov Chain Monte Carlo (MCMC)
}
