% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC.R
\name{fitodeMCMC}
\alias{fitodeMCMC}
\title{Fit ordinary differential equations model using MCMC}
\usage{
fitodeMCMC(model, data, start, tcol = "times", proposal.vcov,
  prior = list(), chains = 1, iter = 2000, burnin = iter/2,
  thin = 1, refresh = max(iter/10, 1), prior.only = FALSE, link,
  fixed = list(), solver.opts = list(method = "rk4"), solver = ode,
  ...)
}
\arguments{
\item{model}{ode model}

\item{data}{data frame with time column and observation column}

\item{start}{named vector of starting parameter values}

\item{tcol}{time column}

\item{prior.only}{(logical) sample from prior distribution only?}

\item{link}{named vector or list of link functions for model parameters}

\item{fixed}{named vector or list of model parameters to fix}

\item{solver.opts}{options for ode integration. See \code{\link{ode}}}

\item{solver}{ode solver}
}
\description{
Fit ordinary differential equations model using MCMC
}
