% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{grad,model.ode-method}
\alias{grad,model.ode-method}
\title{Evaluate the gradients of a model}
\usage{
\S4method{grad}{model.ode}(object, state, par)
}
\arguments{
\item{object}{model.ode object}

\item{state}{state}

\item{par}{parameter values}
}
\description{
Evaluate the gradients of a model
}
