% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{initialize,model.ode-method}
\alias{initialize,model.ode-method}
\title{Initializer for model.ode objects}
\usage{
\S4method{initialize}{model.ode}(.Object, name, model, observation,
  initial, par, link, diffnames, keep_sensitivity = TRUE)
}
\arguments{
\item{.Object}{object}
}
\description{
Initializer for model.ode objects
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{model}}{ode model}

\item{\code{observation}}{observation model}

\item{\code{initial}}{initial values}

\item{\code{par}}{parameters}

\item{\code{link}}{link functions for parameters (log links are used as default)}

\item{\code{diffnames}}{optional character vector specifying the names of a variable for which the consecutive difference needs to be calculated}

\item{\code{keep_sensitivity}}{(logical) maintain the Jacobian as part of the model}
}}

\examples{
SI_model <- new("model.ode",
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    observation = list(
        susceptible ~ dnorm(mean=S, sd=sigma1),
        infected ~ dnorm(mean=I, sd=sigma2)
    ),
    initial = list(
        S ~ N * (1 - i0),
        I ~ N * i0
    ),
    par = c("beta", "gamma", "N", "i0", "sigma1", "sigma2"),
    link = c(i0="logit")
)
}
