% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\docType{methods}
\name{initialize,solution.ode-method}
\alias{initialize,solution.ode-method}
\title{Constructor for solution.ode class}
\usage{
\S4method{initialize}{solution.ode}(
  .Object,
  y,
  times,
  model,
  parms,
  solver.opts = list(method = "rk4"),
  solver = ode
)
}
\arguments{
\item{.Object}{object}

\item{y}{initial values}

\item{times}{time vector}

\item{model}{ode model}

\item{parms}{parameters of the solution}

\item{solver.opts}{options for ode solver}

\item{solver}{ode solver (must take y, times, func, and parms as arguments)}
}
\value{
An object of class ``solution.ode'' as described in \code{\link{solution.ode-class}}.
}
\description{
Constructor for solution.ode class
}
\keyword{internal}
