% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{methods}
\name{initialize,solution.ode-method}
\alias{initialize,solution.ode-method}
\title{the initializer for model.ode}
\usage{
\S4method{initialize}{solution.ode}(.Object, y, times, model, parms,
  ode.opts = list(method = "rk4", hini = 0.1), keep_sensitivity = TRUE)
}
\arguments{
\item{.Object}{object}

\item{y}{initial values}

\item{times}{time vector}

\item{model}{ode model}

\item{parms}{parameters of the solution}

\item{ode.opts}{options for ode integration}

\item{keep_sensitivity}{keep sensitivity equations}
}
\description{
the initializer for model.ode
}

