% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{logLik.sensitivity}
\alias{logLik.sensitivity}
\title{Sensitivity of the likelihood function with respect to parameters}
\usage{
logLik.sensitivity(parms, formula, model, loglik, observation, times = NULL,
  ode.opts = list(method = "rk4", hini = 0.1), returnNLL = TRUE)
}
\description{
Sensitivity of the likelihood function with respect to parameters
}

