% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{logLik.sensitivity}
\alias{logLik.sensitivity}
\title{Calculate the derivative of the log-likelihood function with respect to model parameters}
\usage{
logLik.sensitivity(parms, model, data, solver.opts = list(method =
  "rk4"), solver = ode, return.NLL = TRUE, return.traj = FALSE)
}
\arguments{
\item{parms}{named vector of parameter values}

\item{model}{model.ode object}

\item{data}{data}

\item{solver.opts}{options for the ode solver (see \code{\link{ode}})}

\item{solver}{ode solver}

\item{return.NLL}{(logical) return negative log-likelihood}

\item{return.traj}{(logical) return estimated trajectory}
}
\value{
vector of nll and sensitivity of nll with respect to the parameters
}
\description{
Calculate the derivative of the log-likelihood function with respect to model parameters
}
