% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{ode.sensitivity}
\alias{ode.sensitivity}
\title{Calculate the derivative of the mean expression}
\usage{
ode.sensitivity(
  model,
  parms,
  times,
  solver.opts = list(method = "rk4"),
  solver = ode
)
}
\arguments{
\item{model}{odemodel object}

\item{parms}{named vector of parameter values}

\item{times}{time window for which the model should be solved}

\item{solver.opts}{options for the ode solver (see \code{\link{ode}})}

\item{solver}{ode solver}
}
\description{
Calculate the derivative of an expression with respect to model parameters
using sensitivity equations and chain rule
}
\keyword{internal}
