% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{ode.sensitivity}
\alias{ode.sensitivity}
\title{Calculate sensitivity of the expression with respect to the parameters}
\usage{
ode.sensitivity(expr, model, parms, times, ode.opts = list(method = "rk4",
  hini = 0.1))
}
\arguments{
\item{expr}{expression of model states and parameters}

\item{model}{model.ode object}

\item{parms}{named vector of parameter values}

\item{times}{time window for which the model should be solved}

\item{ode.opts}{options for the ode solver (see \code{\link(ode)})}
}
\description{
Calculate sensitivity of the expression with respect to the parameters
}
\examples{
SI_model <- new("model.ode",
    name = "SI",
    model = list(
        S ~ - beta*S*I/N,
        I ~ beta*S*I/N - gamma*I
    ),
    initial = list(
        S ~ N * (1 - i0),
        I ~ N * i0
    ),
    par= c("beta", "gamma", "N", "i0")
)

ode.sensitivity(expression(gamma*I), SI_model, parms=c(beta=2, gamma=1, N=1e5, i0=1e-4), times=1:10)
}

