% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{ode.sensitivity}
\alias{ode.sensitivity}
\title{Calculate sensitivity of the expression with respect to the parameters}
\usage{
ode.sensitivity(model, parms, times, solver.opts = list(method = "rk4"),
  solver = ode)
}
\arguments{
\item{model}{model.ode object}

\item{parms}{named vector of parameter values}

\item{times}{time window for which the model should be solved}

\item{solver.opts}{options for the ode solver (see \code{\link{ode}})}

\item{expr}{expression to be evaluated}

\item{expr.sensitivity}{partial derivative of expr w.r.t states and parameters}
}
\description{
Calculate sensitivity of the expression with respect to the parameters
}

