% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{ode.solve}
\alias{ode.solve}
\title{solve ode models}
\usage{
ode.solve(
  model,
  times,
  parms,
  y,
  solver.opts = list(method = "rk4"),
  solver = ode
)
}
\arguments{
\item{model}{odemodel object}

\item{times}{time vector}

\item{parms}{named vector of parameter values}

\item{y}{initial values}

\item{solver.opts}{options for ode solver}

\item{solver}{ode solver (must take y, times, func, and parms as arguments)}
}
\value{
An object of class ``solution.ode'' as described in \code{\link{solution.ode-class}}.
}
\description{
solve ode models
}
\keyword{internal}
