% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-class.R, R/model.R
\docType{class}
\name{odemodel-class}
\alias{odemodel-class}
\alias{odemodel}
\title{Class "odemodel" representing ode models}
\usage{
odemodel(...)
}
\description{
Class "odemodel" representing ode models

Create a new odemodel
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{gfun}}{gradient function}

\item{\code{grad}}{list of gradients}

\item{\code{observation}}{list of observation models}

\item{\code{initial}}{list of expressions representing the initial values}

\item{\code{state}}{state variables}

\item{\code{par}}{parameters}

\item{\code{link}}{link functions for parameters (log links are used as default)}

\item{\code{diffnames}}{character vector specifying the names of a variable for which the consecutive difference needs to be calculated}

\item{\code{call}}{original function call}

\item{\code{jacobian.initial}}{Jacobian of initial values with respect to its parameters}

\item{\code{jacobian.state}}{Jacobian with respect to its states}

\item{\code{jacobian.par}}{Jacobian with repsect to its parameters}

\item{\code{loglik}}{list of log-likelihood functions}

\item{\code{expr}}{expressions for true trajectories}

\item{\code{expr.sensitivity}}{sensitivity of the expressions with respect to state variables and parameters}

\item{\code{keep_sensitivity}}{(logical) keep sensitivity equations?}
}}

\keyword{internal}
