% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\name{profile,fitode-method}
\alias{profile,fitode-method}
\title{Profile fitode objects}
\usage{
\S4method{profile}{fitode}(fitted, which = 1:p, alpha = 0.05, trace = FALSE, ...)
}
\arguments{
\item{fitted}{fitted model object}

\item{which}{which parameter(s) to profile? (integer value)}

\item{alpha}{critical level}

\item{trace}{trace progress of computations?}

\item{...}{additional arguments passed to mle2 profiling method}
}
\value{
The log-likelihood profile of the fitode object
}
\description{
Profile fitode objects
}
