% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{profile,fitode-method}
\alias{profile,fitode-method}
\title{Profile fitode objects}
\usage{
\S4method{profile}{fitode}(fitted, which = 1:p, alpha = 0.05,
  trace = FALSE)
}
\arguments{
\item{fitted}{fitted model object}
}
\description{
Profile fitode objects
}
