% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{profile,fitode-method}
\alias{profile,fitode-method}
\title{Profile object}
\usage{
\S4method{profile}{fitode}(fitted, scale = c("original", "fitted"))
}
\arguments{
\item{fitted}{fitted model object}

\item{scale}{scale of parameters}
}
\description{
Profile object
}

