% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{select_model}
\alias{select_model}
\title{Select a log-likelihood model}
\usage{
select_model(
  dist = c("ols", "dnorm", "dnorm2", "dpois", "dnbinom", "dnbinom1", "dgamma",
    "dlnorm")
)
}
\arguments{
\item{dist}{conditional distribution of reported data (dnorm, dnorm2, dpois, dnbinom, dnbinom1, dgamma)}
}
\description{
Select a log-likelihood model
}
\keyword{internal}
