% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode.R
\name{set_link}
\alias{set_link}
\title{Set up link functions for model parameters}
\usage{
set_link(link, par)
}
\arguments{
\item{link}{named list or vector of strings specifying link functions}

\item{par}{(character) model parameters}
}
\value{
list of strings specifying link functions for each model parameter
}
\description{
Set up link functions for model parameters -- for \code{\link{fitode}}
internal usage; assumes log-link by default if link functions are
not specified
}
\seealso{
\code{\link{make.link}}
}
\keyword{internal}
