% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{show,fitode-method}
\alias{show,fitode-method}
\title{Show fitode objects}
\usage{
\S4method{show}{fitode}(object)
}
\arguments{
\item{object}{fitode object}
}
\value{
No return value, called for side effects
}
\description{
Show fitode objects
}
\keyword{internal}
