% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{show,fitode-method}
\alias{show,fitode-method}
\title{Show \code{fitode} object}
\usage{
\S4method{show}{fitode}(object)
}
\arguments{
\item{object}{fitode object}
}
\description{
Show \code{fitode} object
}
