% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{show,fitodeMCMC-method}
\alias{show,fitodeMCMC-method}
\title{Show fitodeMCMC object}
\usage{
\S4method{show}{fitodeMCMC}(object)
}
\arguments{
\item{object}{fitodeMCMC object}
}
\value{
No return value, called for side effects
}
\description{
Show fitodeMCMC object
}
\keyword{internal}
