% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{show,odemodel-method}
\alias{show,odemodel-method}
\title{Show the model}
\usage{
\S4method{show}{odemodel}(object)
}
\arguments{
\item{object}{odemodel object}
}
\value{
No return value, called for side effects
}
\description{
Show the model
}
\keyword{internal}
