% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{solve ode models}
\usage{
solve(model, times, parms, y, ode.opts = list(method = "rk4", hini = 0.1),
  keep_sensitivity = TRUE)
}
\arguments{
\item{model}{model.ode object}

\item{times}{time vector}

\item{parms}{parameters}

\item{y}{initial values}

\item{ode.opts}{options for ode integration}

\item{keep_sensitivity}{keep sensitivity equations}
}
\description{
solve ode models
}

