% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{stdEr,fitode-method}
\alias{stdEr,fitode-method}
\title{Extract standard error from fitode objects}
\usage{
\S4method{stdEr}{fitode}(x, type = c("response", "links"))
}
\arguments{
\item{x}{fitode object}

\item{type}{type of standard error. The default (\code{type=response}) is on the
response scale; this is the scale on which the model parameters are defined.
Alternatively, \code{type=link} can be used to obtain standard errors on the estimated scale.}
}
\value{
The standard error of the fitode object
}
\description{
Calculates standard error by taking the square root of the diagonal matrix
}
