% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{stdEr,fitode-method}
\alias{stdEr,fitode-method}
\title{Calculate standard error}
\usage{
\S4method{stdEr}{fitode}(x, scale = c("original", "fitted"))
}
\arguments{
\item{x}{fitode object}

\item{scale}{scale of parameter to be returned}
}
\description{
Calculate standard error
}

