% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitodeMCMC-methods.R
\docType{methods}
\name{stdEr,fitodeMCMC-method}
\alias{stdEr,fitodeMCMC-method}
\title{Extract standard error from fitodeMCMC objects}
\usage{
\S4method{stdEr}{fitodeMCMC}(x)
}
\arguments{
\item{x}{fitodeMCMC object}
}
\value{
The standard error of the fitodeMCMC object
}
\description{
Calculates standard error by taking the square root of the diagonal of the variance-covariance matrix
}
