% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{summary,fitode-method}
\alias{summary,fitode-method}
\title{Summarize fitode object}
\usage{
\S4method{summary}{fitode}(object)
}
\arguments{
\item{object}{fitode object}
}
\value{
The summary of the fitode object
}
\description{
Summarize fitode objects;
returns estimate, standard error, and confidence intervals
}
