% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Transform,odemodel-method}
\alias{Transform,odemodel-method}
\title{Transform the model}
\usage{
\S4method{Transform}{odemodel}(
  object,
  transforms,
  observation,
  initial,
  par,
  link,
  keep_sensitivity
)
}
\arguments{
\item{object}{odemodel object}

\item{transforms}{list of formulas specifying transformations}

\item{observation}{observation model}

\item{initial}{initial values}

\item{par}{model parameters}

\item{keep_sensitivity}{(logical) maintain the Jacobian as part of the model}
}
\value{
An object of class ``odemodel'' as described in \code{\link{odemodel-class}}.
}
\description{
Transform the model
}
\keyword{internal}
