% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{vcov,fitode-method}
\alias{vcov,fitode-method}
\title{Extract variance-covariance matrix from fitode objects}
\usage{
\S4method{vcov}{fitode}(object, type = c("response", "links"))
}
\arguments{
\item{object}{fitode object}

\item{type}{type of covariance matrix. The default (\code{type=response}) is on the
response scale; this is the scale on which the model parameters are defined.
Alternatively, \code{type=link} can be used to obtain the covariance matrix on the estimated scale.}
}
\value{
The variance-covariance matrix of the fitode object
}
\description{
Extracts variance-covariance matrix (either on response scales or link scales)
}
