% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitode-methods.R
\docType{methods}
\name{vcov,fitode-method}
\alias{vcov,fitode-method}
\title{Extract covariance matrix of a fit}
\usage{
\S4method{vcov}{fitode}(object, type = c("original", "fitted"))
}
\arguments{
\item{object}{fitode object}

\item{type}{type of parameter to be returned}
}
\description{
Extract covariance matrix of a fit
}

